# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Neuralix Dashboard API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /guilds/{guildId}/tickets/config:
    get:
      operationId: getTicketConfig
      tags: [tickets]
      summary: Get ticket system configuration
      parameters:
        - name: guildId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Ticket config
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TicketConfig"
    put:
      operationId: updateTicketConfig
      tags: [tickets]
      summary: Update ticket system configuration
      parameters:
        - name: guildId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TicketConfigInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TicketConfig"
  /guilds/{guildId}/tickets:
    get:
      operationId: getTickets
      tags: [tickets]
      summary: List tickets for a guild
      parameters:
        - name: guildId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Tickets list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Ticket"
  /guilds/{guildId}/tickets/{ticketId}/close:
    post:
      operationId: closeTicket
      tags: [tickets]
      summary: Close a ticket
      parameters:
        - name: guildId
          in: path
          required: true
          schema:
            type: string
        - name: ticketId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Closed

  # ANTIRAID
...
  /guilds/{guildId}/logs/config:
    get:
      operationId: getLogsConfig
      tags: [logs]
      summary: Get logs configuration
      parameters:
        - name: guildId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Logs config
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LogsConfig"
    put:
      operationId: updateLogsConfig
      tags: [logs]
      summary: Update logs configuration
      parameters:
        - name: guildId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LogsConfigInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LogsConfig"
  /guilds/{guildId}/logs:
    get:
      operationId: getGuildLogs
      tags: [logs]
      summary: Get logs for a guild
      parameters:
        - name: guildId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Log entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LogEntry"

  # BLACKLIST
...
  /guilds/{guildId}/premium:
    get:
      operationId: getGuildPremium
      tags: [premium]
      summary: Get guild premium status
      parameters:
        - name: guildId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Premium status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PremiumStatus"
...
  /support/tickets:
    get:
      operationId: getSupportTickets
      tags: [support]
      summary: Get web support tickets
      responses:
        "200":
          description: Support tickets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SupportTicket"
    post:
      operationId: createSupportTicket
      tags: [support]
      summary: Create support ticket
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SupportTicketInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SupportTicket"
  /support/tickets/{id}/messages:
    get:
      operationId: getSupportMessages
      tags: [support]
      summary: Get messages for support ticket
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SupportMessage"
    post:
      operationId: sendSupportMessage
      tags: [support]
      summary: Send message in support ticket
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SupportMessageInput"
      responses:
        "201":
          description: Sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SupportMessage"

  # AI
...
components:
  schemas:
...
    TicketConfig:
      type: object
      properties:
        id:
          type: integer
        guildId:
          type: string
        enabled:
          type: boolean
        categoryId:
          type: ["string", "null"]
        supportRoleId:
          type: ["string", "null"]
        transcriptChannelId:
          type: ["string", "null"]
        maxTicketsPerUser:
          type: integer
        panelMessage:
          type: ["string", "null"]
        panelChannelId:
          type: ["string", "null"]
      required: [id, guildId, enabled, maxTicketsPerUser]
    TicketConfigInput:
      type: object
      properties:
        enabled:
          type: boolean
        categoryId:
          type: ["string", "null"]
        supportRoleId:
          type: ["string", "null"]
        transcriptChannelId:
          type: ["string", "null"]
        maxTicketsPerUser:
          type: integer
        panelMessage:
          type: ["string", "null"]
        panelChannelId:
          type: ["string", "null"]
    Ticket:
      type: object
      properties:
        id:
          type: integer
        guildId:
          type: string
        channelId:
          type: string
        userId:
          type: string
        username:
          type: string
        subject:
          type: string
        status:
          type: string
        claimedBy:
          type: ["string", "null"]
        createdAt:
          type: string
        closedAt:
          type: ["string", "null"]
      required: [id, guildId, channelId, userId, username, subject, status, createdAt]
...
    LogsConfig:
      type: object
      properties:
        id:
          type: integer
        guildId:
          type: string
        enabled:
          type: boolean
        channelId:
          type: ["string", "null"]
        logMembers:
          type: boolean
        logMessages:
          type: boolean
        logRoles:
          type: boolean
        logChannels:
          type: boolean
        logModeration:
          type: boolean
        logSecurity:
          type: boolean
      required: [id, guildId, enabled, logMembers, logMessages, logRoles, logChannels, logModeration, logSecurity]
    LogsConfigInput:
      type: object
      properties:
        enabled:
          type: boolean
        channelId:
          type: ["string", "null"]
        logMembers:
          type: boolean
        logMessages:
          type: boolean
        logRoles:
          type: boolean
        logChannels:
          type: boolean
        logModeration:
          type: boolean
        logSecurity:
          type: boolean
    LogEntry:
      type: object
      properties:
        id:
          type: integer
        guildId:
          type: string
        userId:
          type: ["string", "null"]
        username:
          type: ["string", "null"]
        action:
          type: string
        category:
          type: string
        details:
          type: ["string", "null"]
        moderatorId:
          type: ["string", "null"]
        moderatorName:
          type: ["string", "null"]
        createdAt:
          type: string
      required: [id, guildId, action, category, createdAt]
...
    PremiumStatus:
      type: object
      properties:
        guildId:
          type: string
        active:
          type: boolean
        plan:
          type: ["string", "null"]
        expiresAt:
          type: ["string", "null"]
        features:
          type: array
          items:
            type: string
      required: [guildId, active, features]
...
    SupportTicket:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: string
        username:
          type: string
        subject:
          type: string
        status:
          type: string
        priority:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, userId, username, subject, status, priority, createdAt, updatedAt]
    SupportTicketInput:
      type: object
      properties:
        subject:
          type: string
        message:
          type: string
        priority:
          type: string
      required: [subject, message]
    SupportMessage:
      type: object
      properties:
        id:
          type: integer
        ticketId:
          type: integer
        userId:
          type: string
        username:
          type: string
        avatar:
          type: ["string", "null"]
        content:
          type: string
        isStaff:
          type: boolean
        createdAt:
          type: string
      required: [id, ticketId, userId, username, content, isStaff, createdAt]
    SupportMessageInput:
      type: object
      properties:
        content:
          type: string
      required: [content]
...