# lib/api-zod/src/generated/types/supportMessage.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Neuralix Dashboard API
 * OpenAPI spec version: 0.1.0
 */

export interface SupportMessage {
  id: number;
  ticketId: number;
  userId: string;
  username: string;
  /** @nullable */
  avatar?: string | null;
  content: string;
  isStaff: boolean;
  createdAt: string;
}